app/requirements.txt
Flask
pika
tenacity
pymongo
redis
python-dotenv
scikit-learn
numpy
google-generativeai
joblib